Refactored MotivationToAttackAutomation to use a float instead of an int

+++ core/src/com/unciv/logic/automation/civilization/DeclareWarTargetAutomation.kt
@@ -15,7 +15,7 @@ object DeclareWarTargetAutomation {
      * Chooses a target civilization along with a plan of attack.
      * Note that this doesn't guarantee that we will declare war on them immediatly, or that we will end up declaring war at all.
      */
-    fun chooseDeclareWarTarget(civInfo: Civilization, civAttackMotivations: List<Pair<Civilization, Int>>) {
+    fun chooseDeclareWarTarget(civInfo: Civilization, civAttackMotivations: List<Pair<Civilization, Float>>) {
         val highestValueTargets = civAttackMotivations.sortedByDescending { it.first.getStatForRanking(RankingType.Score) }
 
         for (target in highestValueTargets) {
@@ -27,7 +27,7 @@ object DeclareWarTargetAutomation {
     /**
      * Determines a war plan against this [target] and executes it if able.
      */
-    private fun tryDeclareWarWithPlan(civInfo: Civilization, target: Civilization, motivation: Int): Boolean {
+    private fun tryDeclareWarWithPlan(civInfo: Civilization, target: Civilization, motivation: Float): Boolean {
 
         if (!target.isCityState()) {
             if (motivation > 5 && tryTeamWar(civInfo, target, motivation)) return true
@@ -46,7 +46,7 @@ object DeclareWarTargetAutomation {
      * The safest option for war is to invite a new ally to join the war with us.
      * Together we are stronger and are more likely to take down bigger threats.
      */
-    private fun tryTeamWar(civInfo: Civilization, target: Civilization, motivation: Int): Boolean {
+    private fun tryTeamWar(civInfo: Civilization, target: Civilization, motivation: Float): Boolean {
         val potentialAllies = civInfo.getDiplomacyManager(target)!!.getCommonKnownCivs()
             .filter {
                 it.isMajorCiv()
@@ -74,7 +74,7 @@ object DeclareWarTargetAutomation {
     /**
      * The next safest aproach is to join an existing war on the side of an ally that is already at war with [target].
      */
-    private fun tryJoinWar(civInfo: Civilization, target: Civilization, motivation: Int): Boolean {
+    private fun tryJoinWar(civInfo: Civilization, target: Civilization, motivation: Float): Boolean {
         val potentialAllies = civInfo.getDiplomacyManager(target)!!.getCommonKnownCivs()
             .filter {
                 it.isMajorCiv()
@@ -102,7 +102,7 @@ object DeclareWarTargetAutomation {
     /**
      * Lastly, if our motivation is high enough and we don't have any better plans then lets just declare war.
      */
-    private fun declareWar(civInfo: Civilization, target: Civilization, motivation: Int): Boolean {
+    private fun declareWar(civInfo: Civilization, target: Civilization, motivation: Float): Boolean {
         if (DeclareWarPlanEvaluator.evaluateDeclareWarPlan(civInfo, target, motivation) > 0) {
             civInfo.getDiplomacyManager(target)!!.declareWar()
             return true
@@ -113,7 +113,7 @@ object DeclareWarTargetAutomation {
     /**
      * Slightly safter is to silently plan an invasion and declare war later.
      */
-    private fun prepareWar(civInfo: Civilization, target: Civilization, motivation: Int): Boolean {
+    private fun prepareWar(civInfo: Civilization, target: Civilization, motivation: Float): Boolean {
         // TODO: We use negative values in WaryOf for now so that we aren't adding any extra fields to the save file
         // This will very likely change in the future and we will want to build upon it
         val diploManager = civInfo.getDiplomacyManager(target)!!
